debundle: split purity/mod.rs into domain submodules
The purity classifier lived in a single ~3.9k-line mod.rs. Split it
into cohesive domain submodules behind the same public surface:

  purity_type      Purity enum + constructors
  whitelists       PURE_* builtin/static-call tables
  builtin_calls    new-expression and global/static call classification
  classifier       top-level purity dispatch
  code_graph       owner-graph purity reasons
  plain_data       plain-data initializer / property classification
  ts_enum_iife     TypeScript enum IIFE recognition
  redundant_hints  redundant purity-hint detection

mod.rs is now a thin module index that re-exports the crate-internal
API. No behavior change; all 116 debundle tests pass.
