+++ devinfra/js/debundle/BUILD.bazel
@@ -364,10 +364,17 @@ rust_library(
         "ids.rs",
         "lib.rs",
         "partition.rs",
+        "purity/builtin_calls.rs",
+        "purity/classifier.rs",
         "purity/classifier_tests.rs",
+        "purity/code_graph.rs",
         "purity/graph_purity_tests.rs",
         "purity/mod.rs",
+        "purity/plain_data.rs",
+        "purity/purity_type.rs",
+        "purity/redundant_hints.rs",
         "purity/redundant_hints_tests.rs",
+        "purity/ts_enum_iife.rs",
         "purity/whitelists.rs",
         "reports/mod.rs",
         "reports/schema.rs",
